# api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/brokers": {
      "get": {
        "operationId": "BrokersController_findAll",
        "parameters": [
          {
            "name": "brokerType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "cfd",
                "bond",
                "stock",
                "crypto"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all brokers (paginated, filterable by type)",
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "BrokersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrokerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new broker",
        "tags": [
          "brokers"
        ]
      }
    },
...
    "/brokers/{slug}": {
      "get": {
        "operationId": "BrokersController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single broker by slug (with features, metrics, markets)",
        "tags": [
          "brokers"
        ]
      }
    },
    "/brokers/{id}": {
      "patch": {
        "operationId": "BrokersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrokerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a broker",
        "tags": [
          "brokers"
        ]
      },
      "delete": {
        "operationId": "BrokersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a broker",
        "tags": [
          "brokers"
        ]
      }
    },
...
    "/reviews/broker/{brokerId}": {
      "get": {
        "operationId": "ReviewsController_findByBroker",
        "parameters": [
          {
            "name": "brokerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all reviews for a broker (with reviewer name)",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Woxa Brokers List API",
    "description": "REST API for managing financial brokers",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateBrokerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Exness"
          },
          "slug": {
            "type": "string",
            "example": "exness-a1b2c",
            "description": "Auto-generated from name if omitted"
          },
          "brokerType": {
            "type": "string",
            "enum": [
              "cfd",
              "bond",
              "stock",
              "crypto"
            ],
            "example": "cfd"
          },
          "description": {
            "type": "string",
            "example": "A leading multi-asset broker with deep liquidity."
          },
          "longDescription": {
            "type": "string",
            "example": "Exness is a global multi-asset broker..."
          },
          "website": {
            "type": "string",
            "example": "https://www.exness.com"
          },
          "contactAddress": {
            "type": "string",
            "example": "123 Finance Street, Bangkok, Thailand"
          },
          "contactEmail": {
            "type": "string",
            "example": "support@exness.com"
          },
          "features": {
            "type": "string",
            "description": "JSON-encoded array: [{\"title\":\"...\",\"description\":\"...\",\"sortOrder\":0}]",
            "example": "[{\"title\":\"Zero Commission\",\"description\":\"Trade with zero commission.\"}]"
          }
        },
        "required": [
          "name",
          "brokerType",
          "description",
          "longDescription",
          "website",
          "contactAddress",
          "contactEmail"
        ]
      },
      "UpdateBrokerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Exness"
          },
          "slug": {
            "type": "string",
            "example": "exness-a1b2c",
            "description": "Auto-generated from name if omitted"
          },
          "brokerType": {
            "type": "string",
            "enum": [
              "cfd",
              "bond",
              "stock",
              "crypto"
            ],
            "example": "cfd"
          },
          "description": {
            "type": "string",
            "example": "A leading multi-asset broker with deep liquidity."
          },
          "longDescription": {
            "type": "string",
            "example": "Exness is a global multi-asset broker..."
          },
          "website": {
            "type": "string",
            "example": "https://www.exness.com"
          },
          "contactAddress": {
            "type": "string",
            "example": "123 Finance Street, Bangkok, Thailand"
          },
          "contactEmail": {
            "type": "string",
            "example": "support@exness.com"
          },
          "features": {
            "type": "string",
            "description": "JSON-encoded array: [{\"title\":\"...\",\"description\":\"...\",\"sortOrder\":0}]",
            "example": "[{\"title\":\"Zero Commission\",\"description\":\"Trade with zero commission.\"}]"
          }
        }
      },
...
    }
  }
}